requirements.txt
# Smart Timetable Management System - Requirements
# Only packages actually used in the project

# Core Web Framework
Flask>=3.1.1
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Werkzeug>=3.0.1

# Database
psycopg2-binary>=2.9.9

# Authentication & Security
bcrypt>=4.1.2
PyJWT>=2.8.0

# Data Processing & Analysis
pandas>=2.3.0
numpy>=1.26.0
openpyxl>=3.1.2

# Machine Learning & Optimization
torch>=2.1.0
scikit-learn>=1.3.0
ortools>=9.8.0

# Visualization & Development
plotly>=5.17.0
streamlit>=1.29.0

# Environment Management
python-dotenv>=1.0.0

# HTTP Requests (for API testing)
requests>=2.31.0